\name{bmt}
\alias{bmt}
\docType{data}
\title{
Bone marrow transplant study
}
\description{
Bone marrow transplant study which is widely used in the AFTMC model}
\usage{data(bmt)}
\format{
 There were 46 patients in the allogeneic treatment and 44 patients in the autologous treatment group
  \describe{
    \item{\code{Time}}{ time to event }
    \item{\code{Status}}{ censor indicator, 0 for censored and 1 for uncensored }
    \item{\code{TRT}}{ 1 for autologous treatment group; 0 for allogeneic treatment group}
  }
}

\keyword{datasets}
