\name{coefsmcure}
\alias{coefsmcure}
\title{
Retrieves the estimated coefficients from object
}
\description{
Retrieving coefficients, printing method, and summary mthod for a smcure object. 
}
\usage{
coefsmcure(x,...)
}
\arguments{
  \item{x}{
an object from smcure function
}
 \item{\dots}{
Further arguments to be passed to the printsmcure function.
}
}
