\name{e1684}
\alias{e1684}
\docType{data}
\title{
Eastern Cooperative Oncology Group (ECOG) data
}
\description{
Eastern Cooperative Oncology Group (ECOG) data used for modeling PH semicure model}
\usage{data(e1684)}
\format{
  A data frame with 284 observations on the following 5 variables.
  \describe{
    \item{\code{TRT}}{0=control group, 1=IFN treatment group}
    \item{\code{FAILTIME}}{observed relapse-free time}
    \item{\code{FAILCENS}}{relapse-free censor indicator}
    \item{\code{AGE}}{continuous variable, which is centered to the mean}
    \item{\code{SEX}}{0 for male, 1 fopr female}
  }
}

\keyword{datasets}
