\name{printsmcure}
\alias{printsmcure}
\title{
Print smcure object
}
\description{
Output of smcure object
}
\usage{
printsmcure(x, Var, ...)
}
\arguments{
  \item{x}{
an object of smcure}
  \item{Var}{
If it is TRUE, the program returns standard error by bootstrap method. If set to False, the program only returns estimators of coefficients. By default, Var = TRUE
}
  \item{\dots}{
Further arguments to be passed to the printsmcure function.
}
}

